Last Commit to Master
This commit cleans up the source directory in a way that I think is
minimally acceptable. The convention for testing in Haskell is to have a
copy of the main directory inside the tests directory. I have not lived
with this enough to find it appealing so I'm leaving it out (I'm sure
there are good reasons for it). So I'm keeping my tests in Tests.hs in
the main folder. In another commit I'll probably pull all the
non-compiled files into separate directories and leave the make-ish
stuff in the root of main. I'll add a make file that handles the
standard make cases after that. For now, this works