S4:
This branch is going to be the base branch for the intuitionistic
theorem prover. Right now the algorithm is correct, but inefficient for
the types of theorems I would like to  prove. In particular, it
struggles with equivalance. I suspect that that can be fixed in the
canonicalizer.

It also has the beginnings of a feature for writing proofs out to html.
This would be a nice way to display the proofs.