S4 Feature: Interface
This commit is too big for this branch. I am going to get better about
that.

It includes an update to the way that we generate models that fixes a
bug and but leads to a slowdown. It also includes an interface so that
this program can be called from the command line. Finally, I tried to
get counterexamples for a hypersequent in parallel, but didn't see much
speed up. So I'm commiting that for future reference, but I'll remove it
later.